Define the service
Sec. 15, video 81 of Collaboration and Crawling W/ Go

Ref: https://grpc.io/docs/quickstart/go/

Inside our folder (01-proto), create an echo folder. We're going to
create an echo server. It will just repeat whatever we ask it right back
to us. Create echo.proto.
.proto is the file format for protocol buffer. In this file, we'll
describe a protocol buffer messages and the gRPC service that will be
using these messages with.

The 1st thing will be what version of protocol buffer we're going to be
using. We use the latest version "proto3". If we skip this, protocol
buffer will use "proto2" by default.

Then we define our service. Inside of the service, we can define methods
that can be used. Note that we need "rpc" before our method, just like
"func". In our case, the Echo method takes an EchoRequest and returns an
EchoResponse.

Once we defined the service, we need to define what the EchoRequest and
EchoResponse are. Those are both protocol buffer messages.
Note that the numbers we assigned to the fields in one message have to
be unique and we should not reuse a number, even if we delete it and add
other new fields.