Compile the service
Sec. 15, video 83 of Collaboration and Crawling W/ Go

Now we want to convert the .proto file into a .go package that we can
import. Note that we need to execute the following commands in the
01-proto folder.

We can use the following command to generate the messages:
$ protoc <option> <proto-files>
$ protoc -I echo echo\echo.proto --go_out=echo

-I flag specifies the directory in which to search for imports.
--go_out=OUT_DIR generates Go source file.

The above command will generate the .pb.go file without the echo server
listed in it. It is just the protocol buffer messages part of it
(EchoRequest & EchoResponse).
To generate the echo server, we use the following command:
$ protoc -I echo echo\echo.proto --go_out=plugins=grpc:echo

Now echo.pb.go is generated for us.
Make sure that package name in generated .pb.go file matches the folder
name. In our case, the package name should be "echo".