Set up a gRPC server
Sec. 15, video 85 of Collaboration and Crawling W/ Go

Ref: https://golang.org/pkg/net/#Listen
https://grpc.io/docs/quickstart/go/
https://grpc.io/docs/tutorials/basic/go/

First we copy the 01-proto folder and rename it to 02-server. Then we
create a main.go file inside of it.
In the generated echo.pb.go, we have an EchoServerServer interface (the
name we specify in .proto is EchoServer and it append Server to it), we
need to implement this interface. This is what we will be called
whenever a request is made to this.

In the main function, we have to create a network listener by ourselves
because gRPC server doesn't listen by default on a network port. The
address argument for net.Listen is basically the port.

Initialize a gRPC server and an EchoServer.
Register EchoServer with gRPC server. To do this, we can see that one of
our generated functions, RegisterEchoServerServer, which is responsible
for this.

Now gRPC knows about our service, now we need to actually listen.
We tell our gRPC server to serve on our listener.