Set up a gRPC client
Sec. 15, video 87 of Collaboration and Crawling W/ Go

First we copy the 01-proto folder and rename it to 03-client. Then we
create a main.go file inside of it.
We need to make a connection to the server. We use grpc.Dial to create a
connection to the given target. We also use grpc.WithInsecure to disable
transport security for this client connection.
Note that whenever we got a connection, we need to defer closing that
connection.

Once the gRPC channel is setup, we need a client stub to perform RPCs.
We get this using the NewEchoServerClient method provided in the echo
package we generated from our .proto. It takes in a gRPC connection and
returns an EchoServerClient.
EchoServerClient is the client API for EchoServer service.

Then we can actually call the service method.
Note that in gRPC-Go, RPCs operate in a blocking/synchronous mode, which
means that the RPC call waits for the server to respond, and will either
return a response or an error.