Create a Chat server
Sec. 15, video 90 of Collaboration and Crawling W/ Go

Create a folder 04-chat and in that folder we have a chat folder.
Create a chat.proto file which contains a Chat service.
Note that the Chat method can take in a stream of ChatMessage and return
back a stream of ChatMessage.
So we can make a ChatMessage whenever we want and the server can send
back a ChatMessage when someone else chats whenever.
A ChatMessage is defined as 2 strings, a user and a message.

Compile the chat.proto file to generate chat.pb.go file.
$ protoc -I chat chat\chat.proto --go_out=plugins=grpc:chat

Whenever we've got a stream in the input or output, the function that we
define in Go becomes very different.
Note that on the server side, the Chat function just takes in a
Chat_ChatServer, which just an interface with Send and Recv functions.
It needs to be able to receive a stream of ChatMessage and send a stream
of ChatMessage. Same for the ChatClient.