Create a client
Ref: https://golang.org/pkg/os/#pkg-variables
https://golang.org/pkg/bufio/#Scanner

For the client, we're expecting the user to put into the command line
arguments: the URL to connect to and then the username as a 2nd command
line argument.
os.Args hold the command-line arguments, starting with the program name.
Whenever we have command line arguments, a good very first thing to do
in the main function is to make sure our command line arguments are
valid. In this case we check if we have the right number of command line
arguments.

Create a connection to the given URL.
Then defer closing that connection.
Create a ChatClient.
Initiate that stream on that Chat RPC for both directions.
Since the stream can both be sending and receiving, we use one go
routine to loop over and receive all messages from the stream.
The main go routine will write all the messages to stream with
stream.Send.

We create a channel of empty struct. We're expecting this channel to be
closed before everything ends. That way we can make sure the main
function waits for this go routine (for receiving) to exit before the
main function exits.

There is one thing to note for receiving. If the error we receive back
is an io.EOF, that means the stream is closed in a proper manner. We
should just shut everything down normally. So we close the channel to
indicate this go routine is finished and return out of the go routine
(this go routine is done).

The other half of the program, we're going to print out a statement
basically saying that we're ready to receive messages and explaining how
to exit the program.
We're going to use the bufio.NewScanner to scan because fmt.Scan does
not scan whole lines of code. But bufio.Scanner will let us get whole
lines of code.
bufio stands for buffer area for input and output. Our computer will
often generate all the code the printer needs and then hold it in a
buffer.
os.Stdin is our input on the command line.
As long as there's line to scan, we get the line that just got scanned.
If it is "quit", we close the send direction of the stream.
stream.CloseSend will send out all messages that have not yet been sent
over the stream and close it.
Ultimately, we create a ChatMessage based on the typed in message and
send it.

In the end of the function, we wait for the go routine to finish up.