Create a Chat server
Sec. 15, video 92 of Collaboration and Crawling W/ Go

Ref: https://godoc.org/google.golang.org/grpc#ServerStream
https://blog.golang.org/defer-panic-and-recover
https://gobyexample.com/select

For a Chat program, the server has to receive a message from any
connection then send that message back out to every connection. The
problem is each connection in a gRPC server is running on its own
goroutine, they're all happening simultaneously.

It is safe to have a goroutine calling SendMsg and another goroutine
calling RecvMsg on the same stream at the same time, but it is not safe
to call SendMsg on the same stream in different goroutines.
So we need to setup the thread safety ourselves.

So we made a Connection struct. It holds one connection, which is the
Chat_ChatServer with Send and Recv methods. It has a send channel and a
quit channel.
When we create a Connection with NewConnection function, which takes in
that stream, it will call the start method in another goroutine and
return the Connection.
This is mostly so that we can have the Send method on this Connection
where we can pass in a ChatMessage and this will send the message
through the send channel to the start area.
The start area will actually send the message along that Connection in a
thread safe way. The Send method also ignores any panic in case the send
channel gets closed because the Connection got closed.

The key point is that the only thing that will actually send along the
Connection is the start method in its own goroutine, which we can only
send messages along it through the send channel.

The start method loops forever. Every time it gets a message from the
send channel, it sends it along a Connection. If it ever gets anything
on the quit channel or quit channel closes, it ends the loop and exits
the goroutine. That's why we've got the Close method.

After the Close method closes the send channel, it will cause panic in
Send method. Recover is a built-in function that regains control of a
panicking goroutine. Recover is only useful inside deferred functions.
During normal execution, a call to recover will return nil and have no
other effect. If the current goroutine is panicking, a call to recover
will capture the value given to panic and resume normal execution.

The GetMessages method receives a ChatMessage from the passed in
Connection. We need to send the message to all the connections. This
method takes in a broadcast channel. This is a channel that we are only
allowed to send on for ChatMessages. Some other code elsewhere will be
responsible for make sure that get out to every Send method for all the
connections.
So the GetMessages method receives a message and then it sends  the
message through the broadcast channel.
Note that we're doing this in a new goroutine so that if the other end
of broadcast is not yet ready, it can still start receiving the next
message while the old goroutine waits.
To make it end early, we use select without for loop, which means it
either sends to the broadcast or when quit channel closes. So if the
connection closes, we don't have a bunch of dangling goroutines trying
to send along to broadcast channel.