Handle multiple connections
Sec. 15, video 93 of Collaboration and Crawling W/ Go

Create the overall ChatServer struct which will have a slice of
connections that are currently connected. It has a connLock with type
sync.Mutex. Any time we access the list of connections, we will use the
lock and do the changes and then unlock it afterwards. It also has the
broadcast channel which the connections will send their messages to. So
the ChatServer will receive messages on the broadcast channel and send
them to each of the connections. It also has a server wide quit channel.

The NewChatServer returns the ChatServer. Note that we don't need to
bother with the slice of connections because the append function will
create that if needed. We also don't ever have to actually create a
sync.Mutex because the zero value is valid for it. We run the start
method in a new goroutine.

The start method is the broadcast handler basically. It checks if there
is a message coming in to the broadcast channel. Lock the connLock. Loop
through all the connections. For each connection, send the message in
that goroutine safe way. Note that we do it in another goroutine so that
it doesn't lock up the entire server if some connections being slow.
Then we unlock the connLock.
If the quit channel has things come through or if it's closed, we end
the loop and close this goroutine down. Which we've got a Close method
to kind of abstract that away.

The Chat function implements the prototype of Chat service in chat.proto
. Note that whenever we've got a stream, whether in or out on a
particular method in gRPC, in Go it will change that into a single
argument that's got the Send and Recv methods on it. That's fulfilling
the interface.
It creates that connection and adds it to the left to the slice of
connections. Then it calls GetMessages with the broadcast channel. So
it's able to broadcast back to the server.
When GetMessages returns, that means that connection is done, we remove
the connection from the list of connections.

In the main function, we create a network listener, create a server,
register the server and finally start to serve it.